Refactor home and navbar code
