+++ src/index.js
@@ -1,9 +1,9 @@
 import "./styles/style.css"
 import background from './js/background';
+import navbar from './js/navbar';
 import home from './js/home';
 
-
-
-  const content = document.getElementById('content');
-  content.appendChild(background);
-  content.appendChild(home);
+const content = document.getElementById('content');
+content.appendChild(background);
+content.appendChild(navbar);
+content.appendChild(home);
